+++ jaxrs/src/main/java/org/killbill/billing/jaxrs/json/RolledUpUsageJson.java
@@ -89,7 +89,7 @@ public class RolledUpUsageJson {
         }
 
         public RolledUpUnitJson(final RolledUpUnit input) {
-            this(input.getUnitType(), BigDecimal.valueOf(input.getAmount()) /* FIXME-1469 : API backward compat */);
+            this(input.getUnitType(), input.getAmount());
         }
 
         public String getUnitType() {
